==> R/1.0. Cell area for each of the DGRP lines.r <==
# ----------------------------------------------------------------------------------------------
# Script to calculate cell area twenty DGRP lines
# Data generated by Félix P Leiva, Radboud University
# Created by Felix P Leiva
# ----------------------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls()) 
today<-format(Sys.Date(),"%Y%m%d")
#-----------------------------------------------------------------------------------------------
#get working directory
setwd("C:/Users/Invunche/Dropbox/Radboud University/publicaciones/Thesis/3. Thermal tolerance on big and small cells/datos/size traits")
getwd()# check directory
#-----------------------------------------------------------------------------------------------
#Libraries
library(xlsx)
#-----------------------------------------------------------------------------------------------
#Reading cell size data
datos<-read.xlsx("results on CS available.xlsx", sheetName = "results on CS available")
datos<-subset(datos,datos$INCLUDE=="1")
#-----------------------------------------------------------------------------------------------
#Selecting columns of interest
names(datos)
datos<-datos[c(1:5,7:13)]
#-----------------------------------------------------------------------------------------------
#Rename columns
names(datos)
names(datos)[c(1:12)] = c("stock","genotype","sex","fw","ws","well",
                          "area","trichomes","iod","dm","age","plate")
#-----------------------------------------------------------------------------------------------
#Transforming variables wrongly assigned
str(datos)
datos$stock<-as.factor(datos$stock) 
datos$fw<-as.numeric(datos$fw) 
datos$ws<-as.numeric(datos$ws) 
datos$trichomes<-as.numeric(datos$trichomes) 
datos$area<-as.numeric(datos$area) 
datos$plate<-as.factor(datos$plate) 
#-------------------------------------------------------------------------------
#Average of the three measurements ws 
#Without include IOD
datos2<-aggregate(cbind(ws,fw,trichomes,area)~stock+genotype+sex+age+plate+well,data=datos,mean)#na.rm=TRUE saca las medias descartando los "NA"
#Including IOD
datos3<-aggregate(cbind(ws,fw,trichomes,area,iod)~stock+genotype+sex+age+plate+well,data=datos,mean)#na.rm=TRUE saca las medias descartando los "NA"
#-------------------------------------------------------------------------------
#Calculating cell area for each dataframe
# datos2
datos2$cell.area<-datos2$area/datos2$trichomes
datos2$cell.number<-datos2$ws/datos2$cell.area
# datos3
datos3$cell.area<-datos3$area/datos3$trichomes
datos3$cell.number<-datos3$ws/datos3$cell.area

xtabs(formula = ~sex+stock,data=datos3)
#-------------------------------------------------------------------------------
#Export both data frames
write.csv(datos2,"Individual body traits (no iod) for all DGRP lines.csv")
write.csv(datos3,"Individual body traits for 14 DGRP lines.csv")
#-------------------------------------------------------------------------------
##############################END OF SCRIPT#####################################
#-------------------------------------------------------------------------------
