==> R/3.0. Plotting size traits.r <==
# ------------------------------------------------------------------------------
# Script to generate the following Figures and Tables:
# Figures S1-S2
# Tables S1-S6
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva, Radboud University (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva, Radboud University (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# Cite as:

# Leiva FP, Santos M, Rezende EL & Verberk WCEP. (2021). Paper data and code of
# manuscript: Intraspecific variation on heat tolerance in a model ectotherm:
# the role of oxygen, cell size and body size. Zenodo.
# https://doi.org/10.5281/zenodo.5120028.
# ------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

getwd()# check directory
#-------------------------------------------------------------------------------
#Libraries
library(multcomp)
library(car)
library(lme4)
library(dplyr)
library(MuMIn)
#-------------------------------------------------------------------------------
# load data
datos <- read.csv("../Outputs/2.1.1. Phenotypic traits data of DGRP lines.csv")
#-------------------------------------------------------------------------------
#Checking the structure of data frame
str(datos)
datos$stock    <- as.factor(datos$stock)
datos$plate    <- as.factor(datos$plate)
datos$age      <- as.numeric(datos$age)
datos$well     <- as.factor(datos$well)
datos$sex      <- as.factor(datos$sex)
datos$genotype <- as.factor(datos$genotype)
#-------------------------------------------------------------------------------
#rename some columns
names(datos)
datos <- rename(datos, wing.area = ws)
datos <- rename(datos, fresh.mass = fw)
#-------------------------------------------------------------------------------
# Subset to plot females
female  <- subset(datos, datos$sex=="female")

# Subset to plot males
male    <- subset(datos, datos$sex=="male")
#-------------------------------------------------------------------------------
# Ascending order of cell area (by MEDIAN) for better visualization
female$stock <- reorder(female$stock, female$cell.area.z,FUN=median)

# to check if order is 0K
levels(female$stock)
# "25182" "25198" "25180" "29652" "28191" "28248" "25190" "25192" "28247" "28197" 
# "28141" "28135" "28258" "28153" "28198" "28173" "25203" "25201" "28180" "28196"

# reorder stocks in males as in females to improve the visualization
male$stock <- factor(male$stock, levels = 
                       c("25182", "25198","25180","29652","28191",
                         "28248","25190","25192","28247","28197", 
                         "28141","28135","28258","28153","28198",
                         "28173","25203","25201","28180","28196"))

# Setting same colour for each DGRP line

col.female1=col=c("#CC3333", "#993366", "#666699", "#339999", "#339966", 
                  "#669966", "#666666", "#996699", "#CC6666", "#FF6633",
                  "#FF9900", "#FFCC33", "#FFFF33", "#CC9933", "#996633", 
                  "#c00000", "#CC6699", "#FF99CC", "#CC9999", "#999999")

col.male1=col=c("#CC3333", "#993366", "#666699", "#339999", "#339966",
                "#669966", "#666666", "#996699", "#CC6666", "#FF6633",
                "#FF9900", "#FFCC33", "#FFFF33", "#CC9933", "#996633",
                "#c00000", "#CC6699", "#FF99CC", "#CC9999", "#999999")

#-------------------------------------------------------------------------------
# FIGURE S1
#-------------------------------------------------------------------------------
{
#pdf("../Outputs/3.2.1. Figure S1 Body size traits of 20 DGRP lines.pdf",width = 4.7, height = 7.5, useDingbats = FALSE)
png("../Outputs/3.2.1. Figure S1 Body size traits of 20 DGRP lines.png",width = 4.7, height = 7.5, units = "in", res = 600)
layout(matrix(1:10, ncol = 2, byrow = TRUE))
par(oma = c(3.4, 3.8, 0.2, 3.8), mar=c(0, 0, 0, 0))

#Females
boxplot(fresh.mass ~ stock,data = female, col = col.female1, lwd = 0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(a)", bty = "n")
box()
axis(2,cex.axis = 0.8)
#Males
boxplot(fresh.mass ~ stock, data = male, col=col.male1, lwd = 0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(b)", bty = "n")
box()
axis(4,cex.axis = 0.8)

#Females
boxplot(iod ~ stock, data = female, col = col.female1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(c)",bty = "n")
box()
axis(2,cex.axis = 0.8)
#Males
boxplot(iod~stock, data = male, col = col.male1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(d)",bty = "n")
box()
axis(4,cex.axis = 0.8)

#Females
boxplot(wing.area ~ stock, data = female, col = col.female1, lwd = 0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(e)",bty = "n")
box()
axis(2,cex.axis = 0.8)
#Males
boxplot(wing.area ~ stock, data = male,col = col.male1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(f)",bty = "n")
box()
axis(4,cex.axis = 0.8)

#Females
boxplot(cell.number ~ stock, data = female, col = col.female1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(g)",bty = "n")
box()
axis(2,cex.axis = 0.8)
#Males
boxplot(cell.number ~ stock, data = male, col = col.male1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(h)",bty = "n")
box()
axis(4,cex.axis = 0.8)

boxplot(cell.area ~ stock, data = female, col = col.female1, lwd=0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(i)",bty = "n")
box()
axis(1, at = seq(1, 20, 1), tcl = -0.4,
     cex.axis = 0.6,
     labels = c(25182, 25198, 25180, 29652, 28191, 
                28248, 25190, 25192, 28247, 28197,
                28141, 28135, 28258, 28153, 28198,
                28173, 25203, 25201, 28180, 28196),
     las = 2
)
axis(2,cex.axis = 0.8)

#Males
boxplot(cell.area ~ stock, data = male, col=col.male1, lwd = 0.5, boxwex = 0.7, axes = FALSE)
legend("topleft", legend = "(j)",bty = "n")
axis(1, at = seq(1, 20, 1),tcl = -0.4,
     cex.axis = 0.6,
     labels = c(25182, 25198, 25180, 29652, 28191, 
                28248, 25190, 25192, 28247, 28197,
                28141, 28135, 28258, 28153, 28198,
                28173, 25203, 25201, 28180, 28196),
     las = 2
)
box()
axis(4,cex.axis = 0.8)
# print the overall labels
mtext('DGRP lines', side = 1, outer = TRUE, line = 2.5)
mtext('Body size-traits for females', side = 2, outer = TRUE, line = 2.5)
mtext('Body size-traits for males', side = 4, outer = TRUE, line = 2.5)
dev.off()
}
#-------------------------------------------------------------------------------
# FIGURE S2
#-------------------------------------------------------------------------------
{
pdf("../Outputs/3.2.2. Figure S2 Compare females and males.pdf",width = 3,height = 8,useDingbats = FALSE)
#png("../Outputs/3.2.2. Figure S2 Compare females and males.png",width = 3,height = 8,units = "in",res = 600)
layout(matrix(1:5, ncol=1, byrow=TRUE))
par(mfrow=c(5,1),oma=c(3,3.5,0,0),mar=c(0,4.3,0,0))

boxplot(fresh.mass~sex,data=datos,col=c('#4DAC2660', '#A6761D60'),boxwex=0.5,xaxt='n',cex.axis=0.6,las=1,
        ylab="Fresh mass (mg)")
legend("topleft",legend = "(a)",bty = "n")
box()
boxplot(iod~sex,data=datos,col=c('#4DAC2660', '#A6761D60'),boxwex=0.5,xaxt='n',cex.axis=0.6,las=1,
        ylab=expression(paste("Interocular distance ","(", mu,"m",")")))
legend("topleft",legend = "(b)",bty = "n")
box()
boxplot(wing.area~sex,data=datos,col=c('#4DAC2660', '#A6761D60'),boxwex=0.5,xaxt='n',cex.axis=0.6,las=1,
        ylab=expression(paste("Wing area ","(", mu,"m"^"2",")")))
legend("topleft",legend = "(c)",bty = "n")
box()
boxplot(cell.number~sex,data=datos,col=c('#4DAC2660', '#A6761D60'),boxwex=0.5,xaxt='n',cex.axis=0.6,las=1,
        ylab="Cell number")
legend("topleft",legend = "(d)",bty = "n")
box()
boxplot(cell.area~sex,data=datos,col=c('#4DAC2660', '#A6761D60'),boxwex=0.5,xaxt='n',axt='n',cex.axis=0.6,las=1,
        ylab=expression(paste("Cell area ","(", mu,"m"^"2",")")))
legend("topleft",legend = "(e)",bty = "n")
box()
axis(1,at=c(1,2),labels=c("Females","Males"))

dev.off()
}
#-------------------------------------------------------------------------------
# Do body size traits differ between males and females? (Table S1)
#-------------------------------------------------------------------------------
# FRESH MASS
m1.<-lmer(fresh.mass ~ sex + (1|stock), REML = TRUE, data=datos) # random intercept
m1<-lmer(fresh.mass ~ sex + (1+sex|stock), REML = TRUE, data=datos) # random intercept and slope
anova(m1.,m1);Anova(m1,type="II");AIC(m1.,m1)
logLik(m1);summary(m1);r.squaredGLMM(m1)

# INTERCOCULAR DISTANCE
m2.<-lmer(iod ~ sex + (1|stock), REML = TRUE, data=datos) # random intercept
m2<-lmer(iod ~ sex + (1+sex|stock), REML = TRUE, data=datos) # random intercept and slope
anova(m2.,m2);Anova(m2,type="II");AIC(m2.,m2)
logLik(m2);summary(m2);r.squaredGLMM(m2)

# WING AREA
m3.<-lmer(wing.area ~ sex + (1|stock), REML = TRUE, data=datos) # random intercept
m3<-lmer(wing.area ~ sex + (1+sex|stock), REML = TRUE, data=datos) # random intercept and slope
anova(m3.,m3);Anova(m3,type="II");AIC(m3.,m3)
logLik(m3);summary(m3);r.squaredGLMM(m3)

# CELL NUMBER
m4.<-lmer(cell.number ~ sex + (1|stock), REML = TRUE, data=datos) # random intercept
m4<-lmer(cell.number ~ sex + (1+sex|stock), REML = TRUE, data=datos) # random intercept and slope
anova(m4.,m4);Anova(m4,type="II");AIC(m4.,m4)
logLik(m4);summary(m4);r.squaredGLMM(m4)

# CELL AREA
m5.<-lmer(cell.area ~ sex + (1|stock), REML = TRUE, data=datos) # random intercept
m5<-lmer(cell.area ~ sex + (1+sex|stock), REML = TRUE, data=datos) # random intercept and slope
anova(m5.,m5);Anova(m5,type="II");AIC(m5.,m5)
logLik(m5);summary(m5);r.squaredGLMM(m5)

# Mean of traits by sex 
aggregate(cbind(fresh.mass,iod,wing.area,cell.number,cell.area)~sex,data = datos,mean)

#      sex fresh.mass      iod wing.area cell.number cell.area
# 1 female  1.1764737 475.1735 1062345.2    6391.621  166.6871
# 2   male  0.7607732 441.4189  840264.9    5822.710  144.7141
#-------------------------------------------------------------------------------
# Does body size traits differs across lines?
#-------------------------------------------------------------------------------
# To answer this question, I will fit linear models for each one of the five
# dependent variables, and for each sex. We could do this by using the z-scored
# variables as traits differs in the scale used. Given that all size traits
# differs between sexes, I will analyse the variation separately by sex.

fit.fm.f<-lm(fresh.mass.z~stock,data=female)
summary(fit.fm.f)

#Residuals
res1<-resid(fit.fm.f,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)

# Homocedasticity
boxplot(res1~stock,data=female); 
leveneTest(res1~stock,data=female)

# Tukey test
glht.fm.f<-glht(fit.fm.f,linfct=mcp(stock="Tukey"))#a posteriori test
#summary(glht.fm.f)

cld(glht.fm.f)#assign letters of significant differences

# 25182  25198  25180  29652  28191  28248  25190  25192  28247  28197  28141  28135  28258  28153  28198  28173 
# "cdef"   "bc"   "cd"   "ab"   "dg"  "egh"    "a"    "a"    "h"   "gh"   "cd"   "cd"  "fgh"    "h"    "h"  "egh" 
# 25203  25201  28180  28196 
# "gh"   "cd"  "cde"    "h" 
#-------------------------------------------------------------------------------
# fresh mass males
fit.fm.m<-lm(fresh.mass.z~stock,data=male)
summary(fit.fm.m)

# Residuals...
res1<-resid(fit.fm.m,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)#W = 0.98199, p-value = 0.01358

# Homocedasticity
boxplot(res1~stock,data=male); 
leveneTest(res1~stock,data=male)

# Tukey test
glht.fm.m<-glht(fit.fm.m,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.fm.m)
#cld(glht.fm.m)#assign letters of significant differences

# 25182  25180  25190  29652  25198  28191  28198  28141  28248  28258  25192  28153  28247  28197  28135  28173 
# "bde"   "bc"    "a"    "b"  "bdg"   "dh"   "dh"  "cdh"   "hi" "fghi"  "bde"   "dh"   "dh"  "ghi" "bdef"  "ghi" 
# 25201  28180  25203  28196 
# "eghi"   "bd"   "dh"    "i" 

# Mean and standard deviation of fresh mass per line and sex
grouped <- group_by(datos, stock,sex)
datos.mean<-summarise(grouped, mean=mean(fresh.mass), sd=sd(fresh.mass))

round_df <- function(datos.mean, digits) {
  nums <- vapply(datos.mean, is.numeric, FUN.VALUE = logical(1))
  datos.mean[,nums] <- round(datos.mean[,nums], digits = digits)
  
  (datos.mean)
}

datos.mean<-round_df(datos.mean, digits=3)

#write.csv(datos.mean, "../Outputs/3.1.1. Mean and SD of fresh mass by stock and sex.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# interocular distance females
fit.iod.f<-lm(iod.z~stock,data=female)
summary(fit.iod.f)

# Residuals...
res1<-resid(fit.fm.f,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1) 

# Homocedasticity
boxplot(res1~stock,data=female); 
leveneTest(res1~stock,data=female)

# Tukey test
glht.iod.f<-glht(fit.iod.f,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.iod.f)
cld(glht.iod.f)#assign letters of significant differences
# 25182  25198  25180  29652  28191  28248  25190  25192  28247  28197  28141  28135  28258 
# "eg"   "bd"    "a"   "ab"  "cde"  "cde"   "bd"   "ab"    "g" "cdef"  "cde"   "bc" "cdef" 
# 28153  28198  28173  25203  25201  28180  28196 
# "def"   "fg"  "cde"  "cde" "cdef"   "bd" "cdef"
#-------------------------------------------------------------------------------
# interocular distance males
fit.iod.m<-lm(iod.z~stock,data=male)
summary(fit.iod.m)

# Residuals...
res1<-resid(fit.iod.m,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1) 

# Homocedasticity
boxplot(res1~stock,data=male); 
leveneTest(res1~stock,data=male)

# Tukey test
glht.iod.m<-glht(fit.iod.m,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.iod.m)
cld(glht.iod.m)#assign letters of significant differences

# 25182 25198 25180 29652 28191 28248 25190 25192 28247 28197 28141 28135 28258 28153 28198 
# "ef"  "bd"   "a"  "cd"  "de"  "bd"  "bc"  "cd"   "f"  "cd"  "cd"  "ab"  "de"  "df"  "de" 
# 28173 25203 25201 28180 28196 
# "de"  "de"  "de"  "de"  "bd"

# Mean and standard deviation of interocular distance per line and sex
grouped <- group_by(datos, stock,sex)
datos.mean<-summarise(grouped, mean=mean(iod), sd=sd(iod))

round_df <- function(datos.mean, digits) {
  nums <- vapply(datos.mean, is.numeric, FUN.VALUE = logical(1))
  datos.mean[,nums] <- round(datos.mean[,nums], digits = digits)
  
  (datos.mean)
}

datos.mean<-round_df(datos.mean, digits=3)

#write.csv(datos.mean,"../Outputs/3.1.2. Mean and SD of interocular distance by stock and sex.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# wing area females
fit.wa.f<-lm(wing.area.z~stock,data=female)
summary(fit.wa.f)

# Residuals...
res1<-resid(fit.wa.f,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1) 

# Homocedasticity
boxplot(res1~stock,data=female); 
leveneTest(res1~stock,data=female)

# Tukey test
glht.wa.f<-glht(fit.wa.f,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.wa.f)
#cld(glht.wa.f)#assign letters of significant differences

# 25182  25198  25180  29652  28191  28248  25190  25192  28247  28197  28141  28135  28258  28153  28198  28173 
# "a"  "cde"    "a"    "b"   "ij"  "efg"   "bd"  "efh"   "ef"  "fhi"  "efg"   "bc"   "df"    "j"  "efg"  "fhi" 
# 25203  25201  28180  28196 
# "hj"  "ghj"  "efg" "efhi"
#-------------------------------------------------------------------------------
# wing area males
fit.wa.m<-lm(wing.area.z~stock,data=male)
summary(fit.wa.m)

# Residuals...
res1<-resid(fit.wa.m,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)

# Homocedasticity
boxplot(res1~stock,data=male); 
leveneTest(res1~stock,data=male)

# Tukey test
glht.wa.m<-glht(fit.wa.m,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.wa.m)
#cld(glht.wa.m)#assign letters of significant differences

# 25182 25180 25190 29652 25198 28191 28198 28141 28248 28258 25192 28153 28247 28197 28135 28173 25201 28180 25203 
# "a"   "a"  "bc"   "b" "cef"   "j"  "be" "deg" "hij" "deg"  "ij"  "ij" "bcd" "egh"  "bc" "egi" "fgi"  "bc"  "gi" 
# 28196 
# "fgi

# Mean and standard deviation of interocular distance per line and sex
grouped <- group_by(datos, stock,sex)
datos.mean<-summarise(grouped, mean=mean(wing.area), sd=sd(wing.area))

round_df <- function(datos.mean, digits) {
  nums <- vapply(datos.mean, is.numeric, FUN.VALUE = logical(1))
  datos.mean[,nums] <- round(datos.mean[,nums], digits = digits)
  
  (datos.mean)
}

datos.mean<-round_df(datos.mean, digits=3)

#write.csv(datos.mean,"../Outputs/3.1.3. Mean and SD of wing area by stock and sex.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# cell number females
fit.cn.f<-lm(cell.number.z~stock,data=female)
summary(fit.cn.f)

# Residuals...
res1<-resid(fit.cn.f,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)

# Homocedasticity
boxplot(res1~stock,data=female); 
leveneTest(res1~stock,data=female)

# Tukey test
glht.cn.f<-glht(fit.cn.f,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.cn.f)
# cld(glht.cn.f)#assign letters of significant differences

# 25182  25198  25180  29652  28191  28248  25190  25192  28247  28197  28141  28135  28258  28153  28198  28173 
# "cdef"    "i"   "ab"  "def"    "j"    "i"  "def"   "hi"   "eg"   "gi"  "fgh"   "ac"  "def"   "gi"   "ad" "cdef" 
# 25203  25201  28180  28196 
# "dg" "bcde"    "a"    "a" 
#-------------------------------------------------------------------------------
# cell number males
fit.cn.m<-lm(cell.number.z~stock,data=male)
summary(fit.cn.m)

# Residuals...
res1<-resid(fit.cn.m,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)

# Homocedasticity
boxplot(res1~stock,data=male); 
leveneTest(res1~stock,data=male)

# Tukey test
glht.cn.m<-glht(fit.cn.m,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.cn.m)
#cld(glht.cn.m)#assign letters of significant differences

# 25182 25180 25190 29652 25198 28191 28198 28141 28248 28258 25192 28153 28247 28197 28135 28173 25201 28180 25203 
# "fgh"  "bc" "fgh" "cdf" "gij"   "k" "dfg"  "fi"  "jk" "efg"  "ik" "hij"  "bd" "cdf"  "ab" "cdf" "cdf"   "a" "bde" 
# 28196 
# "ab"

# Mean and standard deviation of cell number per line and sex
grouped <- group_by(datos, stock,sex)
datos.mean<-summarise(grouped, mean=mean(cell.number), sd=sd(cell.number))

round_df <- function(datos.mean, digits) {
  nums <- vapply(datos.mean, is.numeric, FUN.VALUE = logical(1))
  datos.mean[,nums] <- round(datos.mean[,nums], digits = digits)
  
  (datos.mean)
}

datos.mean<-round_df(datos.mean, digits=3)

#write.csv(datos.mean,"../Outputs/3.1.4. Mean and SD of cell number by stock and sex.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# cell area females
fit.ca.f<-lm(cell.area.z~stock,data=female)
summary(fit.ca.f)

# Residuals...
res1<-resid(fit.ca.f,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1) 

# Homocedasticity
boxplot(res1~stock,data=female); 
leveneTest(res1~stock,data=female)

# Tukey test
glht.ca.f<-glht(fit.ca.f,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.ca.f)
#cld(glht.ca.f)#assign letters of significant differences

# 25182 25198 25180 29652 28191 28248 25190 25192 28247 28197 28141 28135 28258 28153 28198 28173 25203 25201 28180 
# "a"   "b"  "bc"  "bd"  "cd" "cde"  "de"  "de"  "ef"  "ef"   "f"   "f"   "f"   "g"  "gh"   "g"  "gi"  "hi"   "i" 
# 28196 
# "j"
# #-----------------------------------------------------------------------------
# cell area males
fit.ca.m<-lm(cell.area.z~stock,data=male)
summary(fit.ca.m)

# Residuals...
res1<-resid(fit.ca.m,type="pearson")

# Normality
qqPlot(res1)
shapiro.test(res1)

# Homocedasticity
boxplot(res1~stock,data=male); 
leveneTest(res1~stock,data=male)

# Tukey test
glht.ca.m<-glht(fit.ca.m,linfct=mcp(stock="Tukey"))#a posteriori test
# summary(glht.ca.m)
#cld(glht.ca.m)#assign letters of significant differences

# 25182 25180 25190 29652 25198 28191 28198 28141 28248 28258 25192 28153 28247 28197 28135 28173 25201 28180 25203 
# "a"   "b"  "bc"  "ce"  "cd" "def"  "ef" "def" "def"  "fg"  "fh" "ghi" "ghi"  "hi" "ghi"   "i"   "i"   "j"   "j" 
# 28196 
# "j"

# Mean and standard deviation of cell area per line and sex
grouped <- group_by(datos, stock, sex)
datos.mean<-summarise(grouped, mean=mean(cell.area), sd=sd(cell.area))

round_df <- function(datos.mean, digits) {
  nums <- vapply(datos.mean, is.numeric, FUN.VALUE = logical(1))
  datos.mean[,nums] <- round(datos.mean[,nums], digits = digits)
  
  (datos.mean)
}

datos.mean<-round_df(datos.mean, digits = 3)

#write.csv(datos.mean,"../Outputs/3.1.5. Mean and SD of cell area by stock and sex.csv",row.names = FALSE)
#-------------------------------------------------------------------------------

#-------------------------------------------------------------------------------
#saving session information with all packages versions for reproducibility
#purposes
sink("../Outputs/3.1.6. Plotting size traits_R_session.txt")
sessionInfo()
sink()
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
