==> R/5.1. Modelling TDT curves under normoxia.r <==
# -------------------------------------------------------------------------------------
# Script to analyse the effect of body size-related traits on survival time of 
# DGRP lines
# Data generated by Felix P Leiva, Radboud University
# Citation: 
# Created by Felix P Leiva
# Modified by Wilco Verberk (20210225)
# Modified by Félix P Leiva (20210226)
# -------------------------------------------------------------------------------------
# Cleaning working space and setting date
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# -------------------------------------------------------------------------------------
# Wilco directory
# setwd("C:/Users/wcepv/Dropbox/VIDI/Félix/Fruitfly paper II/3. Thermal tolerance on big and small cells/3. Thermal tolerance on big and small cells/manuscritos/Submission FE")
# Félix directory
setwd("C:/Users/Invunche/Dropbox/Radboud University/publicaciones/Thesis/3. Thermal tolerance on big and small cells/manuscritos/Submission FE")
# check directory
getwd()
# -------------------------------------------------------------------------------------
#Libraries (we do NOT need to install all these packages)
# library(xlsx)
library(lme4)
library(nlme)
library(car)
library(visreg)
library(MuMIn)
library(AICcmodavg)
library(ggplot2)
library(dplyr)
library(tidyr)
library(plyr)
library(MASS)
# ------------------------------------------------------------------------------
#Load data of normoxia (20 lines)
all.lines.21<-read.csv("Survival time of Drosophila melanogaster and its associated size traits at 21 kPa.csv")

10^(min(all.lines.21$surv.time))#lower range in minutes
10^(max(all.lines.21$surv.time))#upper range in minutes
10^(median(all.lines.21$surv.time))#median in minutes

#Load data of hypoxia (6 lines)
all.lines.hypo<-read.csv("Survival time of Drosophila melanogaster and its associated size traits at two oxygen levels.csv")

10^(min(all.lines.hypo$surv.time))#lower range in minutes
10^(max(all.lines.hypo$surv.time))#upper range in minutes
10^(median(all.lines.hypo$surv.time))#median in minutes

# ------------------------------------------------------------------------------
#Step 1: aggregate data into median values
# ------------------------------------------------------------------------------
d21.median=aggregate(surv.time~test.temp+cell.area+
                       cell.number+stock+genotype+iod+fw+ws+sex,all.lines.21,median)
#count data per stock
count(d21.median$stock)# no data for females of the stock #28180 at 37 and 38°C

#see structure of data
str(d21.median)

#test temp as numeric
d21.median$test.temp<-as.numeric(d21.median$test.temp)

#stock as factor
d21.median$stock<-as.factor(d21.median$stock)

#exploratory plots
plot(surv.time~test.temp,data=all.lines.21)
plot(surv.time~test.temp,data=d21.median)

# ------------------------------------------------------------------------------
# Step 2: Correct traits for effect of sex so that the analysis works with flies 
# that have trait values that are relatively large or small for their sex
# ------------------------------------------------------------------------------
# relative cell area
d21.median$cell.area.rel<-as.numeric(scale(resid(lm(cell.area~sex,data=d21.median))))

# relative cell number
d21.median$cell.number.rel<-as.numeric(scale(resid(lm(cell.number~sex,data=d21.median))))

# relative fresh mass
d21.median$fw.rel<-as.numeric(scale(resid(lm(fw~sex,data=d21.median))))

# relative interocular distance
d21.median$iod.rel<-as.numeric(scale(resid(lm(iod~sex,data=d21.median))))

# relative wing size
d21.median$ws.rel<-as.numeric(scale(resid(lm(ws~sex,data=d21.median))))

# ------------------------------------------------------------------------------
#Step 3: Start with the full model and calculate VIFs
# ------------------------------------------------------------------------------
f21median<-gls(surv.time~
                       scale(test.temp)
               +scale(test.temp)*cell.area.rel
               +scale(test.temp)*cell.number.rel
               +scale(test.temp)*ws.rel
               +scale(test.temp)*fw.rel
               +scale(test.temp)*iod.rel
               +scale(test.temp)*sex
               ,data=d21.median,na.action = na.omit,method = "ML")

vif(f21median)# high VIFs


f21median<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*cell.area.rel
          +scale(test.temp)*cell.number.rel
          #+scale(test.temp)*ws.rel
          +scale(test.temp)*fw.rel
          +scale(test.temp)*iod.rel
          +scale(test.temp)*sex
          ,data=d21.median,na.action = na.omit,method = "ML")

vif(f21median)# VIFs are now 0K!

anova(f21median)
drop1(f21median)
AICc(f21median)#-170.374
# ------------------------------------------------------------------------------
# Step 4: Analyse a list of candidate models (except for wing size) and compare 
# their AICs
# ------------------------------------------------------------------------------
m0<-gls(surv.time~
            scale(test.temp)
          ,data=d21.median,na.action = na.omit,method = "ML")

m1.1<-gls(surv.time~
                scale(test.temp)
                +scale(test.temp)*sex
               ,data=d21.median,na.action = na.omit,method = "ML")

m1.2<-gls(surv.time~
                 scale(test.temp)
               +scale(test.temp)*cell.area.rel
               ,data=d21.median,na.action = na.omit,method = "ML")

m1.3<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

m1.4<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*fw.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

m1.5<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

fit.list.step4 <- list(m0,m1.1,m1.2,m1.3,m1.4,m1.5)

fit.names.step4 <-c("Stress temp",
                 "Stress temp + Stress temp * sex",
                 "Stress temp + Stress temp * cell area",
                 "Stress temp + Stress temp * cell number",
                 "Stress temp + Stress temp * fresh mass",
                 "Stress temp + Stress temp * interocular distance")

#compare by using AICc
fit.step4<-aictab(fit.list.step4,fit.names.step4, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step4

#                                                  K    AICc Delta_AICc AICcWt Cum.Wt    LL
# Stress temp + Stress temp * sex                  5 -167.89       0.00   0.98   0.98 89.14
# Stress temp + Stress temp * fresh mass           5 -160.07       7.81   0.02   1.00 85.23
# Stress temp                                      3 -154.67      13.21   0.00   1.00 80.41
# Stress temp + Stress temp * cell area            5 -152.19      15.70   0.00   1.00 81.29
# Stress temp + Stress temp * cell number          5 -151.40      16.49   0.00   1.00 80.90
# Stress temp + Stress temp * interocular distance 5 -150.45      17.43   0.00   1.00 80.42

# sex and test temperature interaction improve model fit
# ------------------------------------------------------------------------------
# Step 5: Adding to the best model from step 4 the interactive effect of traits 
# and test temperature and then compare their AICs
# ------------------------------------------------------------------------------
m2.1<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel
          ,data=d21.median,na.action = na.omit,method = "ML")
m2.2<-gls(surv.time~
            scale(test.temp)
           +scale(test.temp)*sex
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit,method = "ML")
m2.3<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          ,data=d21.median,na.action = na.omit,method = "ML")
m2.4<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

fit.list.step5 <- list(m1.1,m2.1,m2.2,m2.3,m2.4)

fit.names.step5 <-c("ST + ST * sex",
                 "ST + ST * sex + ST * cell area",
                 "ST + ST * sex + ST * cell number",
                 "ST + ST * sex + ST * fresh mass",
                 "ST + ST * sex + ST * interocular distance")

#compare by using AICc
fit.step5<-aictab(fit.list.step5,fit.names.step5, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step5

#                                           K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST + ST * sex + ST * fresh mass           7 -174.39       0.00   0.94   0.94 94.57
# ST + ST * sex                             5 -167.89       6.50   0.04   0.98 89.14
# ST + ST * sex + ST * cell area            7 -165.63       8.76   0.01   0.99 90.19
# ST + ST * sex + ST * cell number          7 -164.61       9.78   0.01   1.00 89.68
# ST + ST * sex + ST * interocular distance 7 -163.55      10.84   0.00   1.00 89.15

# model that includes sex and relative fresh mass works best
# ------------------------------------------------------------------------------
# Step 6: Adding to the best model from step 5 the interactive effect of traits 
# and test temperature and then compare their AICs
# ------------------------------------------------------------------------------
m3.1<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*cell.area.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

m3.2<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

m3.3<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

fit.list.step6 <- list(m2.3,m3.1,m3.2,m3.3)


fit.names.step6 <-c("ST + ST * sex + ST * fresh mass",
                    "ST + ST * sex + ST * fresh mass + ST * cell area",
                    "ST + ST * sex + ST * fresh mass + ST * cell number",
                    "ST + ST * sex + ST * fresh mass + ST * interocular distance")

#compare by using AICc
fit.step6<-aictab(fit.list.step6,fit.names.step6, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step6

#                                                             K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST + ST * sex + ST * fresh mass                             7 -174.39       0.00   0.41   0.41 94.57
# ST + ST * sex + ST * fresh mass + ST * interocular distance 9 -173.68       0.71   0.28   0.69 96.45
# ST + ST * sex + ST * fresh mass + ST * cell area            9 -173.37       1.02   0.24   0.93 96.29
# ST + ST * sex + ST * fresh mass + ST * cell number          9 -170.84       3.55   0.07   1.00 95.03

# model including sex and relative fresh mass works better, though adding 
# test temp and interocular distance or test temp and cell area is also important
# (DeltaAICc lower than 2).   
# ------------------------------------------------------------------------------
# Lets drop best now the three best model.sel

# (1) Stress temp + Stress temp * sex + Stress temp * fresh mass 
drop1(m2.3,test = "Chisq")
summary(m2.3)

# (2) Stress temp + Stress temp * sex + Stress temp * fresh mass + Stress temp * cell area
drop1(m3.1,test = "Chisq")
summary(m3.1)

# (3) Stress temp + Stress temp * sex + Stress temp * fresh mass + Stress temp * interocular

drop1(m3.3,test = "Chisq")
summary(m3.3)

AIC(m2.3,m3.1,m3.3)

# Only the fixed effect of "scale(test.temp):sex" affects significantly surv time.
# ------------------------------------------------------------------------------
# Step 7: Adding to the best model from step 6 the interactive effect of traits 
# and test temperature and then compare their AICs
# ------------------------------------------------------------------------------
m4.1<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel#given we have specific for cell area, we retain it
          +iod.rel+fw.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

drop1(m4.1,test = "Chisq")
# test temp and cell area interaction and interocular distance are not
# important, so next model exclude this term
m4.2<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +fw.rel
          +cell.area.rel
          ,data=d21.median,na.action = na.omit,method = "ML")
drop1(m4.2,test = "Chisq")

#remove cell area
#
m4.3<-gls(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +fw.rel
          ,data=d21.median,na.action = na.omit,method = "ML")

drop1(m4.3,test = "Chisq")
summary(m4.3)

# Lets now start from a full model
f21median2<-gls(surv.time~
                 scale(test.temp)
               +scale(test.temp)*cell.area.rel
               +scale(test.temp)*cell.number.rel
               #+scale(test.temp)*ws.rel
               +scale(test.temp)*fw.rel
               +scale(test.temp)*iod.rel
               +scale(test.temp)*sex
               ,data=d21.median,na.action = na.omit,method = "ML")
drop1(f21median2)
stepAIC(f21median2)

# Lets write down "more simplified" this model
fmod1<-lm(surv.time ~scale(test.temp) + cell.area.rel + fw.rel + iod.rel +      
            sex + scale(test.temp):cell.area.rel + scale(test.temp):fw.rel +
            scale(test.temp):sex,
          data=d21.median,na.action = na.omit,method = "ML")
AIC(fmod1)#-175.969


m4.3<-lm(surv.time~
           scale(test.temp)
         +scale(test.temp)*sex
         +fw.rel
         ,data=d21.median,na.action = na.omit,method = "ML")
AIC(m4.3)#-176.113

#What is the difference between these two models interm of the AIC?
AIC(fmod1)-AIC(mod1)#0.1440353

#The difference is low if we look at the r squared of both models
summary(fmod1)$r.squared-summary(m4.3)$r.squared
#Given say this and based the most parsimonious model is the "m4.3"

#Just to be sure let now check if there is a three ways interaction 
f21median3<-gls(surv.time~
                 scale(test.temp)
               +scale(test.temp)*cell.area.rel*sex
               ,data=d21.median,na.action = na.omit,method = "ML")

summary(f21median3)# NO three ways interaction

#-------------------------------------------------------------------------------
# TABLE 2 MAIN MANUSCRIPT
#-------------------------------------------------------------------------------
{
m0<-lm(surv.time~ scale(test.temp)
        ,data=d21.median,na.action = na.omit)
summary(m0)

m1.1<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          ,data=d21.median,na.action = na.omit)
summary(m1.1)

m1.2<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*cell.area.rel
          ,data=d21.median,na.action = na.omit)
summary(m1.2)

m1.3<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit)
summary(m1.3)

m1.4<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*fw.rel
          ,data=d21.median,na.action = na.omit)
summary(m1.4)

m1.5<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit)
summary(m1.5)

m2.1<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel
          ,data=d21.median,na.action = na.omit)
summary(m2.1)

m2.2<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit)
summary(m2.2)

m2.3<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          ,data=d21.median,na.action = na.omit)
summary(m2.3)# 

m2.4<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit)
summary(m2.4)

m3.1<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*cell.area.rel
          ,data=d21.median,na.action = na.omit)
summary(m3.1)

m3.2<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*cell.number.rel
          ,data=d21.median,na.action = na.omit)
summary(m3.2)

m3.3<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*fw.rel
          +scale(test.temp)*iod.rel
          ,data=d21.median,na.action = na.omit)
summary(m3.3)

m4.1<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel#given we have specific for cell area, we retain it
          +iod.rel+fw.rel
          ,data=d21.median,na.action = na.omit)
summary(m4.1)

m4.2<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +fw.rel
          +cell.area.rel
          ,data=d21.median,na.action = na.omit)
summary(m4.2)

m4.3<-lm(surv.time~
            scale(test.temp)
          +scale(test.temp)*sex
          +fw.rel
          ,data=d21.median,na.action = na.omit)
summary(m4.3)#MODEL TO PLOT

#                           Estimate Std. Error t value Pr(>|t|)    
# (Intercept)               1.64685    0.01535 107.273  < 2e-16 ***
# scale(test.temp)         -0.42058    0.01532 -27.449  < 2e-16 ***
# sexmale                  -0.03982    0.02157  -1.846 0.066865 .  
# fw.rel                    0.03392    0.01082   3.134 0.002066 ** 
# scale(test.temp):sexmale  0.08550    0.02164   3.951 0.000119 ***

anova(m4.3)

# Response: surv.time
#                       Df  Sum Sq Mean Sq   F value    Pr(>F)    
# scale(test.temp)       1 22.4000 22.4000 1218.4946 < 2.2e-16 ***
# sex                    1  0.0626  0.0626    3.4067 0.0668655 .  
# fw.rel                 1  0.1806  0.1806    9.8234 0.0020660 ** 
# scale(test.temp):sex   1  0.2869  0.2869   15.6085 0.0001186 ***
# Residuals            153  2.8127  0.0184

full<-lm(surv.time~
                 scale(test.temp)
               +scale(test.temp)*cell.area.rel
               +scale(test.temp)*cell.number.rel
               #+scale(test.temp)*ws.rel
               +scale(test.temp)*fw.rel
               +scale(test.temp)*iod.rel
               +scale(test.temp)*sex
               ,data=d21.median,na.action = na.omit)

}
fit.list.table.2 <- list(m0,m1.1,m1.2,m1.3,m1.4,m1.5,
                       m2.1,m2.2,m2.3,m2.4,
                       m3.1,m3.2,m3.3,
                       m4.1,m4.2,m4.3
                       #full
                       )

fit.names.table.2 <-c(
                    "TT",
                    "TT + TT * sex",
                    "TT + TT * CA",
                    "TT + TT * CN",
                    "TT + TT * FM",
                    "TT + TT * IOD",
                    "TT + TT * sex + TT * CA",
                    "TT + TT * sex + TT * CN",
                    "TT + TT * sex + TT * FM",
                    "TT + TT * sex + TT * IOD",
                    "TT + TT * sex + TT * FM + TT * CA",
                    "TT + TT * sex + TT * FM + TT * CN",
                    "TT + TT * sex + TT * FM + TT * IOD",
                    "TT + TT * sex + TT * CA + IOD + FM",
                    "TT + TT * sex + CA + FM",
                    "TT + TT * sex + FM"
                    #"TT + TT * CA + TT * CN + TT * FM + TT * IOD + TT * sex"
                    )

#compare by using AICc
fit.table.2<-aictab(fit.list.table.2,fit.names.table.2, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2

#export table
write.csv(fit.table.2,"Table 2. Model comparison of survival time at 21 kPa.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# # PLOTTING: FIGURE 4
#-------------------------------------------------------------------------------
mydf <- data.frame(d21.median)
plot4<-lm(surv.time~test.temp+test.temp*sex+fw.rel,data=mydf,na.action = na.omit)

{
pdf("Figure 4 Thermal death time curves at 21 kPa.pdf",width = 10,height = 5,useDingbats = FALSE)
png("Figure 4 Thermal death time curves at 21 kPa.png",width = 10,height = 5,units = "in",res = 300)
par(mfrow=c(1,2),tcl=-0.4, family="serif",omi=c(0,0,0,0))
  
# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
  
#by fresh mass
par(mai=c(0.85,0.82,0,0))
visreg(plot4,"test.temp",by="fw.rel",cond=list(sex="male"),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#A6761D70'), lwd=4,lty=c(3,2,1)),
       points=list(col=c('#A6761D70'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
 
quantile(d21.median$fw.rel) 
# Add a legend
legend("topleft", legend=c("(a)"), cex=1.2,bty="n")
legend("topright", legend=c("small FM", "medium FM","large FM"),
       lty = c(3,2,1),lwd=3,bty="n",cex=1.2, 
       col=c('#A6761D70'))

# By sex
par(mai=c(0.85,0.6,0,0.2))
visreg(plot4,"test.temp",by="sex",
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670', '#A6761D70'),lwd=4),yaxt="n",ylab="Survival time",
       points=list(col=c('#4DAC2670', '#A6761D70'),pch=16,cex=1.4),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)

# Add a legend
legend("topleft", legend=c("(b)"), cex=1.2,bty="n")
legend("topright", legend=c("female", "male"),
       lwd=4,bty="n",
       col=c('#4DAC2670', '#A6761D70'),cex=1.2)

# close device
dev.off()
}
