==> R/5.2. Modelling TDT curves under two oxygen conditions.r <==
# -------------------------------------------------------------------------------------
# Script to analyse the effect of body size-related traits on survival time of 
# DGRP lines at two oxygen conditions
# Citation:
# Data generated by Félix P Leiva, Radboud University
# Created by Félix P Leiva 20210309
# -------------------------------------------------------------------------------------
# Cleaning working space and setting date
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# -------------------------------------------------------------------------------------
# Wilco directory
# setwd("C:/Users/wcepv/Dropbox/VIDI/Félix/Fruitfly paper II/3. Thermal tolerance on big and small cells/3. Thermal tolerance on big and small cells/manuscritos/Submission FE")
# Félix directory
setwd("C:/Users/Invunche/Dropbox/Radboud University/publicaciones/Thesis/3. Thermal tolerance on big and small cells/manuscritos/Submission FE")
# check directory
getwd()
# -------------------------------------------------------------------------------------
#Libraries (we do NOT need to install all these packages)
# library(xlsx)
library(lme4)
library(nlme)
library(car)
library(visreg)
library(MuMIn)
library(AICcmodavg)
library(ggplot2)
library(dplyr)
library(tidyr)
library(plyr)
library(MASS)
# ------------------------------------------------------------------------------
#Load data of hypoxia (6 lines)
all.lines.hypo<-read.csv("Survival time of Drosophila melanogaster and its associated size traits at two oxygen levels.csv")
# ------------------------------------------------------------------------------
#Step 1: aggregate data into median values
# ------------------------------------------------------------------------------
dhypo.median=aggregate(surv.time~test.temp+cell.area+
                       cell.number+stock+test.oxygen+iod+fw+ws+sex,all.lines.hypo,median)
#count data per stock
count(dhypo.median$stock)

#see structure of data
str(dhypo.median)

#test temp as numeric
dhypo.median$test.temp<-as.numeric(dhypo.median$test.temp)

#test oxygen as numeric
dhypo.median$test.oxygen<-as.numeric(dhypo.median$test.oxygen)

#stock as factor
dhypo.median$stock<-as.factor(dhypo.median$stock)

#exploratory plots
plot(surv.time~test.temp,data=all.lines.hypo)
plot(surv.time~test.temp,data=dhypo.median)

# ------------------------------------------------------------------------------
# Step 2: Correct traits for effect of sex so that the analysis works with flies 
# that have trait values that are relatively large or small for their sex
# ------------------------------------------------------------------------------
# relative cell area
dhypo.median$cell.area.rel<-as.numeric(scale(resid(lm(cell.area~sex,data=dhypo.median))))

# relative cell number
dhypo.median$cell.number.rel<-as.numeric(scale(resid(lm(cell.number~sex,data=dhypo.median))))

# relative fresh mass
dhypo.median$fw.rel<-as.numeric(scale(resid(lm(fw~sex,data=dhypo.median))))

# relative interocular distance
dhypo.median$iod.rel<-as.numeric(scale(resid(lm(iod~sex,data=dhypo.median))))

# relative wing size
dhypo.median$ws.rel<-as.numeric(scale(resid(lm(ws~sex,data=dhypo.median))))

# ------------------------------------------------------------------------------
#Step 3: Start with the full model and calculate VIFs
# ------------------------------------------------------------------------------
fhypomedian<-gls(surv.time~
                       scale(test.temp)
               +scale(test.temp)*scale(test.oxygen)
               +scale(test.temp)*cell.area.rel*scale(test.oxygen)
               +scale(test.temp)*cell.number.rel*scale(test.oxygen)
               +scale(test.temp)*ws.rel*scale(test.oxygen)
               +scale(test.temp)*fw.rel*scale(test.oxygen)
               +scale(test.temp)*iod.rel*scale(test.oxygen)
               +scale(test.temp)*sex*scale(test.oxygen)
               ,data=dhypo.median,na.action = na.omit,method = "ML")

vif(fhypomedian)# high VIFs

fhypomedian<-gls(surv.time~
                 scale(test.temp)
               +scale(test.oxygen)
               +scale(test.oxygen)*scale(test.temp)
               +scale(test.temp)*cell.area.rel*scale(test.oxygen)
               +scale(test.temp)*cell.number.rel*scale(test.oxygen)
               #+scale(test.temp)*ws.rel*scale(test.oxygen)
               +scale(test.temp)*fw.rel*scale(test.oxygen)
               +scale(test.temp)*iod.rel*scale(test.oxygen)
               +scale(test.temp)*sex*scale(test.oxygen)
               ,data=dhypo.median,na.action = na.omit,method = "ML")

vif(fhypomedian)# VIFs are now 0K!

anova(fhypomedian)
drop1(fhypomedian)
stepAIC(fhypomedian)
AICc(fhypomedian)#-77.44816
# ------------------------------------------------------------------------------
# Step 4: Analyse a list of candidate models (except for wing size) and compare 
# their AICs
# ------------------------------------------------------------------------------
m0<-gls(surv.time~
            scale(test.temp)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m0.1<-gls(surv.time~
          scale(test.temp)*scale(test.oxygen)
        ,data=dhypo.median,na.action = na.omit,method = "ML")

m1.1<-gls(surv.time~
                scale(test.temp)*scale(test.oxygen)
                +scale(test.temp)*sex*scale(test.oxygen)
               ,data=dhypo.median,na.action = na.omit,method = "ML")

m1.2<-gls(surv.time~
                 scale(test.temp)*scale(test.oxygen)
               +scale(test.temp)*cell.area.rel*scale(test.oxygen)
               ,data=dhypo.median,na.action = na.omit,method = "ML")

m1.3<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*cell.number.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m1.4<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*fw.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m1.5<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*iod.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

fit.list.step4 <- list(m0,m0.1,m1.1,m1.2,m1.3,m1.4,m1.5)

fit.names.step4 <-c("ST",
                 "ST * O2",  
                 "ST * O2 + ST * sex * O2",
                 "ST * O2 + ST * cell area * O2",
                 "ST * O2 + ST * cell number * O2",
                 "ST * O2 + ST * FM * O2",
                 "ST * O2 + ST * IOD * O2")

#compare by using AICc
fit.step4<-aictab(fit.list.step4,fit.names.step4, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step4

#                                 K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST * O2 + ST * sex * O2         9 -102.26       0.00   0.50   0.50 61.23
# ST * O2 + ST * cell area * O2   9 -101.85       0.41   0.41   0.91 61.02
# ST * O2                         5  -98.32       3.94   0.07   0.98 54.51
# ST * O2 + ST * IOD * O2         9  -93.80       8.46   0.01   0.99 57.00
# ST * O2 + ST * FM * O2          9  -92.58       9.68   0.00   1.00 56.39
# ST * O2 + ST * cell number * O2 9  -92.47       9.80   0.00   1.00 56.33
# ST                              3  -83.93      18.33   0.00   1.00 45.10

# sex and test temperature and oxygen interaction improve model fit. Important to
# highlight is the effect of test temperature and oxygen interaction and cell area. 
# It seems quite important also!!!
# ------------------------------------------------------------------------------
# Step 5: Adding to the best model from step 4 the interactive effect of traits 
# and test temperature and then compare their AICs
# ------------------------------------------------------------------------------
m2.1<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m2.2<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.number.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m2.3<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*fw.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m2.4<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*iod.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

fit.list.step5 <- list(m1.1,m2.1,m2.2,m2.3,m2.4)

fit.names.step5 <-c("ST * O2 + ST * sex * O2",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Cell area",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Cell number",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Fresh mass",
                    "ST * O2 + ST * sex * O2 + ST * O2 * IOD"
                    )

#compare by using AICc
fit.step5<-aictab(fit.list.step5,fit.names.step5, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step5

#                                                  K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area   13 -107.07       0.00   0.90   0.90 68.87
# ST * O2 + ST * sex * O2                          9 -102.26       4.81   0.08   0.99 61.23
# ST * O2 + ST * sex * O2 + ST * O2 * IOD         13  -97.55       9.52   0.01   0.99 64.11
# ST * O2 + ST * sex * O2 + ST * O2 * Fresh mass  13  -96.25      10.82   0.00   1.00 63.46
# ST * O2 + ST * sex * O2 + ST * O2 * Cell number 13  -95.62      11.45   0.00   1.00 63.14

# ------------------------------------------------------------------------------
# Step 6: Adding to the best model from step 5 the interactive effect of traits 
# and test temperature and then compare their AICs
# ------------------------------------------------------------------------------
m3.1<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*fw.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m3.2<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*cell.number.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

m3.3<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*iod.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit,method = "ML")

fit.list.step6 <- list(m2.1,m3.1,m3.2,m3.3)


fit.names.step6 <-c("ST * O2 + ST * sex * O2 + ST * O2 * Cell area",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * FM",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * CN",
                    "ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * IOD")

#compare by using AICc
fit.step6<-aictab(fit.list.step6,fit.names.step6, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step6

#                                                               K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area                 13 -107.07       0.00   0.86   0.86 68.87
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * IOD 17 -103.09       3.98   0.12   0.98 72.68
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * FM  17  -98.28       8.79   0.01   0.99 70.27
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area + ST * O2 * CN  17  -97.32       9.75   0.01   1.00 69.79

# model including sex and relative fresh mass works better, though adding 
# test temp and interocular distance or test temp and cell area is also important
# (DeltaAICc lower than 2).   
# ------------------------------------------------------------------------------
# Lets drop the best model

# (1) ST * O2 + ST * sex * O2 + ST * O2 * Cell area 
drop1(m2.1,test = "Chisq")
summary(m2.1)
anova(m2.1)

# Cell size interacts with stress temp and also with oxygen. The most significant
# effect os due to test.temp and test.temp and oxygen interaction. Similarly as in
# the experiment under normoxia , there is a significant effects of the interaction
# between sex and test temp. 
# ------------------------------------------------------------------------------
# Step 7: Lets now exclude the non significant variables
# ------------------------------------------------------------------------------
#exclude the three-ways interaction

m4.1<-gls(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel
          +scale(test.oxygen)*cell.area.rel
          ,data=dhypo.median,na.action = na.omit,method = "ML")

drop1(m4.1,test = "Chisq")
summary(m4.1)
anova(m4.1)

# account only for the interactive effect
m4.2<-gls(surv.time~
            scale(test.temp):scale(test.oxygen)
          +scale(test.temp)*sex
          +scale(test.temp):cell.area.rel
          +scale(test.oxygen):cell.area.rel
          ,data=dhypo.median,na.action = na.omit,method = "ML")
drop1(m4.2,test = "Chisq")
summary(m4.2)
anova(m4.2)

fit.list.step7 <- list(m2.1,m4.1,m4.2)

fit.names.step7 <-c("ST * O2 + ST * sex * O2 + ST * O2 * Cell area",
                    "ST * O2 + ST * Sex + ST * Cell area + O2 * Cell area",
                    "ST : O2 + ST * Sex + ST : Cell area + O2 : Cell area")

#compare by using AICc
fit.step7<-aictab(fit.list.step7,fit.names.step7, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.step7

#                                                       K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST : O2 + ST * Sex + ST : Cell area + O2 : Cell area  8 -113.42       0.00   0.58   0.58 65.58
# ST * O2 + ST * Sex + ST * Cell area + O2 * Cell area 10 -112.67       0.75   0.40   0.98 67.69
# ST * O2 + ST * sex * O2 + ST * O2 * Cell area        13 -107.07       6.35   0.02   1.00 68.87
#-------------------------------------------------------------------------------
# TABLE 4 MAIN MANUSCRIPT
#-------------------------------------------------------------------------------
{
m0<-lm(surv.time~
            scale(test.temp)
          ,data=dhypo.median,na.action = na.omit)
summary(m0)  
m0.1<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m0.1)
  
m1.1<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*sex*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m1.1)  

m1.2<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*cell.area.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m1.2)
  
m1.3<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*cell.number.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m1.3)  

m1.4<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*fw.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m1.4)  

m1.5<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*iod.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m1.5)  

m2.1<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*sex*scale(test.oxygen)
            +scale(test.temp)*cell.area.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m2.1)  

m2.2<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*sex*scale(test.oxygen)
            +scale(test.temp)*cell.number.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m2.2)  

m2.3<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*sex*scale(test.oxygen)
            +scale(test.temp)*fw.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m2.3)  

m2.4<-lm(surv.time~
              scale(test.temp)*scale(test.oxygen)
            +scale(test.temp)*sex*scale(test.oxygen)
            +scale(test.temp)*iod.rel*scale(test.oxygen)
            ,data=dhypo.median,na.action = na.omit)
summary(m2.4)  

m3.1<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*fw.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit)
summary(m3.1)

m3.2<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*cell.number.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit)
summary(m3.2)

m3.3<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex*scale(test.oxygen)
          +scale(test.temp)*cell.area.rel*scale(test.oxygen)
          +scale(test.temp)*iod.rel*scale(test.oxygen)
          ,data=dhypo.median,na.action = na.omit)
summary(m3.3)

m4.1<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel
          +scale(test.oxygen)*cell.area.rel
          ,data=dhypo.median,na.action = na.omit)
summary(m4.1)

m4.2<-lm(surv.time~
            scale(test.temp):scale(test.oxygen)
          +scale(test.temp)*sex
          +scale(test.temp):cell.area.rel
          +scale(test.oxygen):cell.area.rel
          ,data=dhypo.median,na.action = na.omit)
summary(m4.2)#BEST MODEL TO PLOT

}
fit.list.table.4 <- list(m0,m0.1,m1.1,m1.2,m1.3,m1.4,m1.5,
                         m2.1,m2.2,m2.3,m2.4,
                         m3.1,m3.2,m3.3,
                         m4.1)

fit.names.table.4 <-c("ST",
                      "ST * O2",  
                      "ST * O2 + ST * sex * O2",
                      "ST * O2 + ST * CA * O2",
                      "ST * O2 + ST * CN * O2",
                      "ST * O2 + ST * FM * O2",
                      "ST * O2 + ST * IOD * O2",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA",
                      "ST * O2 + ST * sex * O2 + ST * O2 * IOD",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * FM",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * CN",
                      "ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * IOD",
                      "ST * O2 + ST * sex + ST * CA + O2 * CA"
                      )

#compare by using AICc
fit.table.4<-aictab(fit.list.table.4,fit.names.table.4, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.4

# Model selection based on AICc:
#   
#                                                         K    AICc Delta_AICc AICcWt Cum.Wt    LL
# ST * O2 + ST * sex + ST * CA + O2 * CA                 10 -112.67       0.00   0.92   0.92 67.69
# ST * O2 + ST * sex * O2 + ST * O2 * CA                 13 -107.07       5.60   0.06   0.98 68.87
# ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * IOD 17 -103.09       9.58   0.01   0.99 72.68
# ST * O2 + ST * sex * O2                                 9 -102.26      10.41   0.01   0.99 61.23
# ST * O2 + ST * CA * O2                                  9 -101.85      10.81   0.00   1.00 61.02
# ST * O2                                                 5  -98.32      14.35   0.00   1.00 54.51
# ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * FM  17  -98.28      14.39   0.00   1.00 70.27
# ST * O2 + ST * sex * O2 + ST * O2 * IOD                13  -97.55      15.12   0.00   1.00 64.11
# ST * O2 + ST * sex * O2 + ST * O2 * CA + ST * O2 * CN  17  -97.32      15.35   0.00   1.00 69.79
# ST * O2 + ST * sex * O2 + ST * O2 * CA                 13  -96.25      16.42   0.00   1.00 63.46
# ST * O2 + ST * sex * O2 + ST * O2 * CA                 13  -95.62      17.05   0.00   1.00 63.14
# ST * O2 + ST * IOD * O2                                 9  -93.80      18.86   0.00   1.00 57.00
# ST * O2 + ST * FM * O2                                  9  -92.58      20.08   0.00   1.00 56.39
# ST * O2 + ST * CN * O2                                  9  -92.47      20.20   0.00   1.00 56.33
# ST                                                      3  -83.93      28.73   0.00   1.00 45.10

#export table
write.csv(fit.table.4,"Table 4. Model comparison of survival time at two kPa.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# # PLOTTING: FIGURE 5
#-------------------------------------------------------------------------------
mydf <- data.frame(dhypo.median)
plot5<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)
          +scale(test.temp)*sex
          +scale(test.temp)*cell.area.rel
          +scale(test.oxygen)*cell.area.rel
         ,data=dhypo.median,na.action = na.omit)
summary(plot5)#BEST MODEL TO PLOT
anova(plot5)
quantile(dhypo.median$cell.area.rel, prob=c(0,0.25,0.5,0.75,1))

{
pdf("Figure 5 Thermal death time curves at two kPa.pdf",width = 10,height = 11,useDingbats = FALSE)
png("Figure 5 Thermal death time curves at two kPa.png",width = 10,height = 10,units = "in",res = 300)
par(mfrow=c(2,2),tcl=-0.4, family="serif",omi=c(0,0,0,0))

# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))

# By sex
par(mai=c(0.6,0.82,0,0))
visreg(plot5,"test.temp",by="sex",cond=list(test.oxygen=21),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670', '#A6761D70'),lwd=4),yaxt="n",
       points=list(col=c('#4DAC2670', '#A6761D70'),pch=16,cex=1.4),
       cex.axis=1.2,xlab="",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(a)"), cex=1.4,bty="n")
legend("topright", legend=c("female", "male"),
       lwd=4,bty="n",
       col=c('#4DAC2670', '#A6761D70'),cex=1.2)

#by oxygen
par(mai=c(0.6,0.6,0,0.2))
visreg(plot5,"test.temp",by="test.oxygen",
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c("#E66101", "#00AFBB"), lwd=4),
       points=list(col=c('#A6761D70'),pch=16,cex=1.4),
       cex.axis=1.2,xlab="",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(b)"), cex=1.4,bty="n")
legend("topright", legend=c("10 kPa", "21 kPa"),
       lwd=4,bty="n",
       col=c("#E66101", "#00AFBB"),cex=1.2)

#by cell area at hypoxia
par(mai=c(0.85,0.82,0,0))
visreg(plot5,"test.temp",by="cell.area.rel",cond=list(test.oxygen=10),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c('#E66101'), lwd=4,lty=c(3,2,1)),
       points=list(col=c('#4DAC2670'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))


#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(c) females tested at 10 kPa"), cex=1.4,bty="n")
legend("topright", legend=c("small CA", "medium CA","large CA"),
       lty = c(3,2,1),lwd=4,bty="n",cex=1.2, 
       col=c('#E66101'))

#by cell area at normoxia
par(mai=c(0.85,0.6,0,0.2))
visreg(plot5,"test.temp",by="cell.area.rel",cond=list(test.oxygen=21),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c('#00AFBB'), lwd=4,lty=c(3,2,1)),
       points=list(col=c('#4DAC2670'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,cex=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(d) females tested at 21 kPa"), cex=1.4,bty="n")
legend("topright", legend=c("small CA", "medium CA","large CA"),
       lty = c(3,2,1),lwd=4,bty="n",cex=1.2, 
       col=c('#00AFBB'))

# close device
dev.off()
}

################################################################################
#calculate quantile of relative cell area
#for females
q10.f<-with(subset(dhypo.median,sex=="female"),quantile(cell.area.rel,p=0.1, na.rm=T))
q10.f# -1.563393

q90.f<-with(subset(dhypo.median,sex=="female"),quantile(cell.area.rel,p=0.9, na.rm=T))
q90.f# 1.361103

qmed.f<-with(subset(dhypo.median,sex=="female"),quantile(cell.area.rel,p=0.5, na.rm=T))
qmed.f# 0.1568448

# for males
q10.m<-with(subset(dhypo.median,sex=="male"),quantile(cell.area.rel,p=0.1, na.rm=T))
q10.m# -1.577031

q90.m<-with(subset(dhypo.median,sex=="male"),quantile(cell.area.rel,p=0.9, na.rm=T))
q90.m# 1.165957

qmed.m<-with(subset(dhypo.median,sex=="male"),quantile(cell.area.rel,p=0.5, na.rm=T))
qmed.m# 0.2951262

#fit a four-wys interaction model to interpret better the output of tables 3 and
#4
plot.test<-lm(surv.time~
            scale(test.temp)*scale(test.oxygen)*cell.area.rel*sex
          ,data=dhypo.median,na.action = na.omit)
#exploratory plot
{
pdf("Figure to interprete results from Table 3 and 4.pdf",width = 10,height = 10,useDingbats = FALSE)
png("Figure to interprete results from Table 3 and 4.png",width = 10,height = 10,units = "in",res = 300)

par(mfrow=c(2,2),tcl=-0.4, family="serif",omi=c(0,0,0,0))

#plot for small cells at 10 kPa
par(mai=c(0,0.5,0.5,0))
visreg(plot.test,"test.temp",by="sex",cond=list(test.oxygen=10,cell.area.rel=-1.5),
      overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
      ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
      line=list(col=c('#4DAC2670','#A6761D70'),lwd=4),
      cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
      xlim=c(35.6,39.4),ylim=c(0,3))
# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "small cells at 10 kPa")
legend("topright", legend=c("Females","Males"),
       lwd=3,bty="n",cex=1, 
       col=c('#4DAC2670','#A6761D70'))

#plot for large cells at 10 kPa
par(mai=c(0,0.5,0.5,0))
visreg(plot.test,"test.temp",by="sex",cond=list(test.oxygen=10,cell.area.rel=1.2),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670','#A6761D70'),lwd=4),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "large cells at 10 kPa")
legend("topright", legend=c("Females","Males"),
       lwd=3,bty="n",cex=1, 
       col=c('#4DAC2670','#A6761D70'))

#plot for small cells at 21 kPa
par(mai=c(0.5,0.5,0.5,0))
visreg(plot.test,"test.temp",by="sex",cond=list(test.oxygen=21,cell.area.rel=-1.5),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670','#A6761D70'),lwd=4),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "small cells at 21 kPa")
legend("topright", legend=c("Females","Males"),
       lwd=3,bty="n",cex=1, 
       col=c('#4DAC2670','#A6761D70'))

#plot for large cells under normoxia
par(mai=c(0.5,0.5,0.5,0))
visreg(plot.test,"test.temp",by="sex",cond=list(test.oxygen=21,cell.area.rel=1.2),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670','#A6761D70'),lwd=4),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "large cells at 21 kPa")
legend("topright", legend=c("Females","Males"),
       lwd=3,bty="n",cex=1, 
       col=c('#4DAC2670','#A6761D70'))
dev.off()
}
################################################################################
{
pdf("Figure to interprete results from Table 3 and 4 (option2).pdf",width = 10,height = 5,useDingbats = FALSE)
png("Figure to interprete results from Table 3 and 4 (option2).png",width = 10,height = 5,units = "in",res = 300)

par(mfrow=c(1,2),tcl=-0.4, family="serif",omi=c(0,0,0,0))
  
#only males at 10 kPa
par(mai=c(0,0.5,0.5,0))
visreg(plot.test,"test.temp",by="cell.area.rel",cond=list(test.oxygen=10),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#A6761D70'), lwd=4,lty=c(3,2,1)),
       #points=list(col=c('#A6761D70'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "males in hypoxia")
legend("topright", legend=c("small CA", "medium CA","large CA"),
       lty = c(3,2,1),lwd=3,bty="n",cex=1.2, 
       col=c('#A6761D70'))

#only males at 21 kPa
par(mai=c(0,0.5,0.5,0))
v1<-visreg(plot.test,"test.temp",by="cell.area.rel",cond=list(test.oxygen=21),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=FALSE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#A6761D70'), lwd=4,lty=c(3,2,1)),
       #points=list(col=c('#A6761D70'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (°C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)
#add grid for better visualization
abline(h=(seq(0,7,0.2)), col="lightgray", lty="dotted")
# Add a legend
legend("topleft",legend = "males in normoxia")
legend("topright", legend=c("small CA", "medium CA","large CA"),
       lty = c(3,2,1),lwd=3,bty="n",cex=1.2, 
       col=c('#A6761D70'))

dev.off()
}
################################################################################
########################### END OF SCRIPT ######################################
################################################################################
# Tests for Mauro and Enrico

#see structure of data
str(all.lines.hypo)

#test temp as numeric
all.lines.hypo$test.temp<-as.numeric(all.lines.hypo$test.temp)

#test oxygen as numeric
all.lines.hypo$test.oxygen<-as.numeric(all.lines.hypo$test.oxygen)

#stock as factor
all.lines.hypo$stock<-as.factor(all.lines.hypo$stock)

#ozygen as factor
all.lines.hypo$test.oxygen<-as.factor(all.lines.hypo$test.oxygen)

#exploratory plots
plot(surv.time~test.temp,data=all.lines.hypo)


# relative cell area
all.lines.hypo$cell.area.rel<-as.numeric(scale(resid(lm(cell.area~sex,data=all.lines.hypo))))

# relative cell number
all.lines.hypo$cell.number.rel<-as.numeric(scale(resid(lm(cell.number~sex,data=all.lines.hypo))))

# relative fresh mass
all.lines.hypo$fw.rel<-as.numeric(scale(resid(lm(fw~sex,data=all.lines.hypo))))

# relative interocular distance
all.lines.hypo$iod.rel<-as.numeric(scale(resid(lm(iod~sex,data=all.lines.hypo))))

# relative wing size
all.lines.hypo$ws.rel<-as.numeric(scale(resid(lm(ws~sex,data=all.lines.hypo))))

#models
lme.test<-lmer(surv.time~test.temp*test.oxygen+(1+test.oxygen|stock),REML=TRUE, data=all.lines.hypo)
summary(lme.test)
anova(lme.test)
library(smatr)
mslope_test<-with(all.lines.hypo,sma(surv.time~test.temp*stock,method="MA"))
mslope_test
plot(mslope_test,type = "l")
