==> R/7.0. Plotting Figure 2 and Figure 3.r <==
# -------------------------------------------------------------------------------------
# Script to plot the effect of fresh mass, sex, oxygen and cell area on survival time of 
# D. melanogaster under normoxic and hypoxic conditions
# -------------------------------------------------------------------------------------
# Data generated by F?lix P Leiva, Radboud University (e-mail: felixpleiva@gmail.com)
# Script created by F?lix P Leiva, Radboud University (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# Cite as:

# Leiva FP, Santos M, Rezende EL & Verberk WCEP. (2021). Paper data and code of
# manuscript: Intraspecific variation on heat tolerance in a model ectotherm:
# the role of oxygen, cell size and body size. Zenodo.
# https://doi.org/10.5281/zenodo.5120028.
# ------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

getwd()# check directory
# -------------------------------------------------------------------------------------
#Libraries
library(nlme)
library(car)
library(visreg)
library(MuMIn)
library(AICcmodavg)
library(MASS)
# ------------------------------------------------------------------------------
#Load data of normoxia (20 lines)
all.lines.21 <- read.csv("../Outputs/4.1.1. Survival time of Drosophila melanogaster and its associated size traits at 21 kPa.csv")

10^(min(all.lines.21$surv.time)) # lower range of survival time in minutes
10^(max(all.lines.21$surv.time)) # upper range of survival time in minutes
10^(median(all.lines.21$surv.time)) # median of survival time in minutes

#Load data of hypoxia (6 lines)
all.lines.hypo<-read.csv("../Outputs/4.1.2. Survival time of Drosophila melanogaster and its associated size traits at two oxygen levels.csv")

10^(min(all.lines.hypo$surv.time))#lower range in minutes
10^(max(all.lines.hypo$surv.time))#upper range in minutes
10^(median(all.lines.hypo$surv.time))#median in minutes
# ------------------------------------------------------------------------------
#Step 1.1: aggregate data into median values normoxia
# ------------------------------------------------------------------------------
d21.median=aggregate(surv.time~test.temp+cell.area+
                       cell.number+stock+genotype+iod+fw+ws+sex,all.lines.21,median)

#see structure of data
str(d21.median)

#test temp as numeric
d21.median$test.temp<-as.numeric(d21.median$test.temp)

#stock as factor
d21.median$stock<-as.factor(d21.median$stock)

# ------------------------------------------------------------------------------
# Step 1.2: Correct traits for effect of sex so that the analysis works with flies 
# that have trait values that are relatively large or small for their sex (normoxia)
# ------------------------------------------------------------------------------
# relative cell area
d21.median$cell.area.rel<-as.numeric(scale(resid(lm(cell.area~sex,data=d21.median))))

# relative cell number
d21.median$cell.number.rel<-as.numeric(scale(resid(lm(cell.number~sex,data=d21.median))))

# relative fresh mass
d21.median$fw.rel<-as.numeric(scale(resid(lm(fw~sex,data=d21.median))))

# relative interocular distance
d21.median$iod.rel<-as.numeric(scale(resid(lm(iod~sex,data=d21.median))))

# relative wing size
d21.median$ws.rel<-as.numeric(scale(resid(lm(ws~sex,data=d21.median))))


# ------------------------------------------------------------------------------
#Step 2.1: aggregate data into median values
# ------------------------------------------------------------------------------
dhypo.median=aggregate(surv.time~test.temp+cell.area+
                         cell.number+stock+test.oxygen+iod+fw+ws+sex,all.lines.hypo,median)

#see structure of data
str(dhypo.median)

# transform structure of data
dhypo.median$test.temp<-as.numeric(dhypo.median$test.temp)
dhypo.median$test.oxygen<-as.numeric(dhypo.median$test.oxygen)
dhypo.median$stock<-as.factor(dhypo.median$stock)
dhypo.median$sex<-as.factor(dhypo.median$sex)

# ------------------------------------------------------------------------------
# Step 2.2: Correct traits for effect of sex so that the analysis works with flies 
# that have trait values that are relatively large or small for their sex
# ------------------------------------------------------------------------------
# relative cell area
dhypo.median$cell.area.rel<-as.numeric(scale(resid(lm(cell.area~sex,data=dhypo.median))))

# relative cell number
dhypo.median$cell.number.rel<-as.numeric(scale(resid(lm(cell.number~sex,data=dhypo.median))))

# relative fresh mass
dhypo.median$fw.rel<-as.numeric(scale(resid(lm(fw~sex,data=dhypo.median))))

# relative interocular distance
dhypo.median$iod.rel<-as.numeric(scale(resid(lm(iod~sex,data=dhypo.median))))

# relative wing size
dhypo.median$ws.rel<-as.numeric(scale(resid(lm(ws~sex,data=dhypo.median))))

#-------------------------------------------------------------------------------
# BEST MODELS FOR PLOTTING
#-------------------------------------------------------------------------------
#best model for survival time measured under normoxia
# mydf <- data.frame(d21.median)
fit_5 <- lm(surv.time ~ scale(test.temp) +
              scale(test.temp) * sex +
              scale(test.temp) * fw.rel,
            data = d21.median, na.action = na.omit)
summary(fit_5)

#best model for survival time measured under hypoxia
# mydf <- data.frame(dhypo.median)
hyp_5 <- lm(surv.time ~ scale(test.temp) +
                    scale(test.temp) * scale(test.oxygen) * sex +
                    scale(test.temp) * scale(test.oxygen) * cell.area.rel,
            data = dhypo.median, na.action = na.omit)
summary(hyp_5)

#quatiles of cell area
quantile(dhypo.median$cell.area.rel, prob=c(0.25,0.5,0.75))

#-------------------------------------------------------------------------------
# FIGURE 2
#-------------------------------------------------------------------------------
{
pdf("../Outputs/7.2.1. Figure 2.pdf", width = 10,height = 3, useDingbats = FALSE)
#png("../Outputs/7.2.1. Figure 2.png",width = 10,height = 3,units = "in",res = 1200)
par(mfrow=c(1,3),tcl=-0.4, family="serif",omi=c(0,0,0,0))

# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))

#by fresh mass
par(mai=c(0.6,0.6,0,0))
visreg(fit_5,"test.temp",by="fw.rel",cond=list(sex="male"),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#A6761D70'), lwd=4,lty=c(3,2,1)),
       points=list(col=c('#55555560'),pch=16,cex=c(0.8,1.2,1.6)),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)

quantile(d21.median$fw.rel) 
# Add a legend
legend("topleft", legend=c("(a)"), cex=1.6,bty="n")
legend("topright", legend=c("small FM", "medium FM","large FM"),
       lty = c(3,2,1),lwd=2,bty="n",cex=1.2, 
       col=c('#A6761D70'))

# By sex
par(mai=c(0.6,0.6,0,0))
visreg(fit_5,"test.temp",by="sex",
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.2, yaxt="n",rug=FALSE,
       line=list(col=c('#4DAC2670', '#A6761D70'),lwd=4),yaxt="n",
       points=list(col=c('#55555560', '#55555560'),pch=16,cex=1.4),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))
#update axis
axis(2, at=xs, labels=labs, cex.axis=1.2)

# Add a legend
legend("topleft", legend=c("(b)"), cex=1.6,bty="n")
legend("topright", legend=c("female", "male"),
       lwd=2,bty="n",
       col=c('#4DAC2670', '#A6761D70'),cex=1.2)

#by oxygen
par(mai=c(0.6,0.6,0,0))
visreg(hyp_5,"test.temp",by="test.oxygen",
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c("#E66101", "#1c4ad6"), lwd=4),
       points=list(col=c('#55555560'),pch=16,cex=1.4),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(c)"), cex=1.6,bty="n")
legend("topright", legend=c("10 kPa", "21 kPa"),
       lwd=2,bty="n",
       col=c("#E66101", "#1c4ad6"),cex=1.2)

# close device
dev.off()
}
#-------------------------------------------------------------------------------
# FIGURE 3
#-------------------------------------------------------------------------------
{
#pdf("../Outputs/7.2.2. Figure 3.pdf", width = 10,height = 3, useDingbats = FALSE)
png("../Outputs/7.2.2. Figure 3.png",width = 10,height = 3,units = "in",res = 1200)
par(mfrow=c(1,3),tcl=-0.4, family="serif",omi=c(0,0,0,0))
        
# labels
labs= c("1","2","5","10","30","120","420")
xs= log10(c(1,2,5,10,30,60*2,60*7))

# by oxygen for small cells
par(mai=c(0.6,0.6,0,0))
visreg(hyp_5,"test.temp",by="test.oxygen",cond=list(cell.area.rel=-0.9449488),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c("#E66101", "#1c4ad6"), lwd=4),
       points=list(col=c('#55555560'),pch=16,cex=c(1.4)),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(a) small cells"), cex=1.6,bty="n")
legend("topright", legend=c("10 kPa", "21 kPa"),
       lwd=2,bty="n",
       col=c("#E66101", "#1c4ad6"),cex=1.2)


# by oxygen for medium cells
par(mai=c(0.6,0.6,0,0))
visreg(hyp_5,"test.temp",by="test.oxygen",cond=list(cell.area.rel=0.2951262),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c("#E66101", "#1c4ad6"), lwd=4),
       points=list(col=c('#55555560'),pch=16,cex=c(1.4)),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(b) medium cells"), cex=1.6,bty="n")
legend("topright", legend=c("10 kPa", "21 kPa"),
       lwd=2,bty="n",
       col=c("#E66101", "#1c4ad6"),cex=1.2)


# by oxygen for large cells
par(mai=c(0.6,0.6,0,0))
visreg(hyp_5,"test.temp",by="test.oxygen",cond=list(cell.area.rel=0.7404176),
       overlay =TRUE,band=FALSE,jitter=0.1,partial=TRUE,legend=FALSE,
       ylab="Time (min)",cex.lab=1.4, yaxt="n",rug=FALSE,
       line=list(col=c("#E66101", "#1c4ad6"), lwd=4),
       points=list(col=c('#55555560'),pch=16,cex=c(1.4)),
       cex.axis=1.2,xlab="Test temperature (?C)",cex.lab=1.4,
       xlim=c(35.6,39.4),ylim=c(0,3))

#update axis
axis(2, at=xs, labels=labs, cex.axis=1.4)

# Add a legend
legend("topleft", legend=c("(c) large cells"), cex=1.6,bty="n")
legend("topright", legend=c("10 kPa", "21 kPa"),
       lwd=2,bty="n",
       col=c("#E66101", "#1c4ad6"),cex=1.2)

# close device
dev.off()
}

#-------------------------------------------------------------------------------
#saving session information with all packages versions for reproducibility
#purposes
sink("../Outputs/7.1.1. TDT curves under normoxia_R_session.txt")
sessionInfo()
sink()
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
